Move org summary into it's own file to improve readability.
